requirements.txt
adjustText==1.3.0
ag-ui-protocol==0.1.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
bleach==6.2.0
boto3==1.40.21
botocore==1.40.21
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chromadb==1.0.20
click==8.1.8
cohere==5.17.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.0.0
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dlai-grader==1.22.2
durationpy==0.10
eval_type_backport==0.2.2
executing==2.2.0
fastavro==1.12.0
fastjsonschema==2.21.2
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.1
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
genai-prices==0.0.24
google-auth==2.40.3
google-genai==1.32.0
googleapis-common-protos==1.70.0
griffe==1.13.0
groq==0.31.0
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
invoke==2.2.0
IProgress==0.4
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
jupytext==1.17.3
kagglehub==0.3.12
kiwisolver==1.4.9
kubernetes==33.1.0
lark==1.2.2
logfire-api==4.3.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mcp==1.13.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
mistralai==1.9.9
mistune==3.1.4
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nexus-rpc==1.1.0
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.3.2
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.101.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pillow==11.3.0
platformdirs==4.3.8
posthog==5.4.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-ai==0.8.1
pydantic-ai-slim==0.8.1
pydantic-evals==0.8.1
pydantic-graph==0.8.1
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.3
sympy==1.14.0
tabulate==0.9.0
temporalio==1.16.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
together==1.5.25
tokenizers==0.21.4
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.4
typer==0.15.4
types-protobuf==6.30.2.20250822
types-python-dateutil==2.9.0.20250822
types-requests==2.32.4.20250809
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uuid7==0.1.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
xgboost==3.0.4
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
